error message to users if they've uploaded a shapefile that isn't a polygon (#268)
* error message to users if they've uploaded a shapefile that isn't a polygon

https://eaflood.atlassian.net/browse/LTFRI-1116
