+++ node/risk-admin/server/public/static/js/create-comment.js
@@ -26,11 +26,12 @@ fileInput.addEventListener('change', function (e) {
   fetch('/shp2json/' + type, {
     method: 'post',
     body: formData
-  }).then(function (response) {
+  }).then(async function (response) {
     spinner.style.display = 'none'
 
     if (!response.ok) {
-      throw new Error(response.statusText)
+      const result = await response.json()
+      throw new Error(result.message)
     }
 
     return response
@@ -109,7 +110,7 @@ fileInput.addEventListener('change', function (e) {
 
     if (jsonFileData.features.length > 1) {
       commentMap(jsonFileData, 'map', capabilities, 'The map below shows all geometries contained within the shapefile')
-    } 
+    }
 
     // Add char count for the text areas
     const textareas = document.querySelectorAll('textarea')
@@ -122,17 +123,17 @@ fileInput.addEventListener('change', function (e) {
       })
     })
 
-    function updateRemainingChars(textarea, remainingCharsText) {
+    function updateRemainingChars (textarea, remainingCharsText) {
       const maxLength = parseInt(textarea.getAttribute('maxLength'))
       remainingCharsText.innerHTML = maxLength - textarea.value.length
     }
-    
+
     return jsonFileData
   }).then(function (jsonFileData) {
     // Construct the form data checking for any changes made by user in fields ready for payload
     const commentForm = document.getElementById('comment-form')
 
-    function handleFormSubmit(event) {
+    function handleFormSubmit (event) {
       event.preventDefault()
       const eventFormData = new FormData(event.target)
 
@@ -144,9 +145,9 @@ fileInput.addEventListener('change', function (e) {
         const riskOverrideValue = eventFormData.get(`override_${index}-risk`)
         const riskReportType = eventFormData.get(`features_${index}_properties_report_type`)
         const addCommentRadio = eventFormData.get(`add_holding_comment_${index}`)
-        
-        if (jsonFileData.name !== eventFormData.get(`name`)) {
-          jsonFileData.name = eventFormData.get(`name`)
+
+        if (jsonFileData.name !== eventFormData.get('name')) {
+          jsonFileData.name = eventFormData.get('name')
         }
         if (jsonFileData.features[index].properties.end !== eventFormData.get(`features_${index}_properties_end`)) {
           jsonFileData.features[index].properties.end = eventFormData.get(`features_${index}_properties_end`)
@@ -165,14 +166,13 @@ fileInput.addEventListener('change', function (e) {
           }
           if (addCommentRadio === 'No') {
             jsonFileData.features[index].properties.commentText = 'No'
-            jsonFileData.features[index].properties.info = ""
+            jsonFileData.features[index].properties.info = ''
           } else {
             jsonFileData.features[index].properties.commentText = 'Yes'
           }
         } else {
           jsonFileData.features[index].properties.info = riskReportType
         }
-
       })
     }
 
@@ -199,6 +199,6 @@ fileInput.addEventListener('change', function (e) {
     })
   }).catch(function (err) {
     console.error(err)
-    window.alert('Invalid shapefile')
+    window.alert('Invalid shapefile: ' + err.message)
   })
 })
